Expand mapsec IDs over 0xFF by using u16 variables

+++ src/pokenav_match_call_data.c
@@ -40,7 +40,7 @@ struct MatchCallStructCommon {
 
 struct MatchCallStructNPC {
     u8 type;
-    u8 mapSec;
+    u16 mapSec;
     u16 flag;
     const u8 *desc;
     const u8 *name;
